requirements.txt
django
django-method-override
pillow
six
pytz
netaddr
deprecation
django_tables2
python-dateutil
pytimeparse
pycryptodome
django-ajax-selects
django-summernote
django-ajax-selects
django-maintenance-mode
djangorestframework
markdown
django-json-widget

#Optional
django-impersonate
django-guardian
